backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
pydantic==2.6.4
pytest==8.1.1
pytest-asyncio==0.23.5
databases==0.9.0
alembic==1.13.1
python-jose>=3.3.0
